+++ requirements.txt
@@ -5,3 +5,4 @@ py-cord>=2.6.1
 pyyaml>=6.0.1
 toml>=0.10.2
 flask[async]>=2.0.0
+tenacity=8.5.0
